+++ contracts/services/osmosis-gamm-lper/Cargo.toml
@@ -20,12 +20,12 @@ cw-ownable = { workspace = true }
 valence-macros = { workspace = true }
 valence-service-utils = { workspace = true }
 valence-service-base = { workspace = true }
-cosmwasm-std-polytone = { package = "cosmwasm-std", version = "1.5.7" }
 osmosis-std = { workspace = true }
 valence-account-utils = { workspace = true }
 valence-osmosis-utils = { workspace = true }
 
 [dev-dependencies]
+cosmwasm-std-polytone = { package = "cosmwasm-std", version = "1.5.7" }
 cw20 = { workspace = true }
 osmosis-test-tube = { workspace = true }
 valence-osmosis-utils = { workspace = true, features = ["testing"] }
