+++ local-interchaintest/src/utils/manager.rs
@@ -25,7 +25,7 @@ pub fn setup_manager(test_ctx: &mut TestContext) -> Result<(), Box<dyn Error>> {
     let artifacts_dir = format!("{}/artifacts", env::current_dir()?.to_str().unwrap());
     let chain_infos = get_data_from_log();
     let mut gc = get_global_config();
-    gc.chains = chain_infos.clone();
+    gc.chains.clone_from(&chain_infos);
 
     let mut uploader = test_ctx.build_tx_upload_contracts();
     uploader.with_chain_name(NEUTRON_CHAIN_NAME);
